Organize code
- Refactored and re-organized code into new hierarchies.
