+++ app/src/main/AndroidManifest.xml
@@ -14,11 +14,11 @@
         android:theme="@style/Theme.Remember"
         tools:targetApi="31">
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.HomeActivity"
             android:exported="false"
             android:theme="@style/Theme.Remember" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
