+++ app/src/main/res/layout/launcher_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity"
+    tools:context=".ui.HomeActivity"
     android:background="@color/black_coral">
 
     <TextView
